requirements.txt
Django==1.11.2
djangorestframework==3.6.3
mysqlclient==1.3.12
django-rest-knox
Pillow